+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- feat: add metadata.metrics.config.extraProcessors [#2724]
+
+[#2724]: https://github.com/SumoLogic/sumologic-kubernetes-collection/pull/2724
+[Unreleased]: https://github.com/SumoLogic/sumologic-kubernetes-collection/compare/v3.0.0-beta.0...main
+
 ## [v3.0.0-beta.1]
 
 ### Changed